+++ Client/qtTeamTalk/utilui.h
@@ -148,5 +148,5 @@ bool restoreWindowPosition(const QString& setting, QWidget* widget);
 QStringList extractLanguages();
 bool switchLanguage(const QString& language);
 QString getFormattedDateTime(QString originalDateTime, QString format);
-QString getFormattedFileSize(int filesize);
+QString getFormattedFileSize(qint64 filesize);
 #endif
